Hoist canvas code into it's own component
